Add CLI arguments --max-iter and --threshold
These CLI arguments control how long the convergence loop should take

Replace NetworkX methods to get TPG n° nodes and n° edges with two
properties.
This is the first step into removing NetworkX from the project because
it can not handle the HUGE size of the graph we are going to build (thus
untill now we built it in a wrong way). In we simply did not consider
the Preference nodes not connected to any user; but this is conceptually
wrong and causes the TPG to be possibly disconnected into several
sub-graphs.