Add -m/--member and -g/--group-size CLI arguments
This commit introduces process level parallelism since the thread one
alone usually can not saturate the available cores.

We can for example split the workload between two processes with:
    ./grank.py input_file --member 0 --group-size 2 -o shared_file
    ./grank.py input_file --member 1 --group-size 2 -o shared_file
and collect the whole results in the shared_file.

Simultaneous write access to the shared file is protected with a lock on
it. Actually that kind of file locking primitives is available only on
unix based systems.