Compute vector representing a document
Corpus.vector_of(doc_id) now returns a vector (actually a sparse matrix
but with a unitary dimension) where each cell is the normalized weight
of a term (i.e. token) in the document.

Some caching has been used for:
- the whole number of tokens in a document
- the maximum frequency of terms in a document
- the inverse document frequency (IDF) of a term in the corpus
- the denormalized TF-IDF of a term in a document
- the cosine normalized TF-IDF of a term in a document
Otherwise the computation of a document vector was really long.